Initial building/snapping prototype
Temporary controls: Middle mouse button to toggle between build mode, mouse wheel to change which part. Only a non-drivable car at the moment. For right now, parts will only snap to nodes that have their name as a tag in the node's name, and only in the same position and orientation as the node.