Logo files added to Resources
The Logo files are inn both PNG and SVG format and on transparent backgrounds. The large one is configured for use on 4K UHD (3840x2160) not DCI 4K (4096x2160). This can be modified if we decide to use the Motion Picture Standard (DCI 4K) or stay with the TV and Consumer Media Standard (4K UHD).