Updated folder structure and packed rigged, animated models.
I have set up a couple of models for testing. The Vanguard model has considerably more anims with it. I selected what I thought would be a useful collection.  All models should have a locomotion package for basic and extended movement, but Vanguard has a much larger selection of deforms.

There are scripts for converting the fbx files to 3DS Max bipeds in the Characters folder.

I have also setup a folder structure that I will be using moving forward to try and keep things tidy, as it is becoming difficult to keep track of with some stuff.  The structure is taken from indie-Pixel. I will, as time permits, start getting the editor extensions and panels built that will use the folder structure exclusively.